pyproject.toml
[tool.poetry]
name = "my-finance"
version = "0.1.0"
description = "Personal Finance Web App"
authors = ["rafael-lacerda <67244039+rafael-lacerda@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.1.2"
alembic = "^1.11.1"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.19"
pydantic = "^2.0.3"
uvicorn = "^0.23.1"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.2"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-check = "^2.2.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.flake8]
ignore = ['E231', 'E241']
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 120
count = true

[tool.pylint]
max-line-length = 120
disable = "C0103,C0114,C0115,C0116,W0511"
ignored-modules = "os,sys,time"
